requirements.txt
pdfreader
cryptography
pycryptodome
pyasn1
pyasn1-modules